mm_modem: Move common types to own file
